+++ app/src/main/java/cord/eoeo/momentwo/di/FriendModule.kt
@@ -1,12 +1,12 @@
 package cord.eoeo.momentwo.di
 
-import cord.eoeo.momentwo.data.MomentwoDatabase
+import cord.eoeo.momentwo.core.database.MomentwoDatabase
 import cord.eoeo.momentwo.data.friend.FriendDataSource
 import cord.eoeo.momentwo.data.friend.FriendRemoteMediator
 import cord.eoeo.momentwo.data.friend.FriendRepositoryImpl
-import cord.eoeo.momentwo.data.friend.local.FriendDao
+import cord.eoeo.momentwo.core.database.FriendDao
 import cord.eoeo.momentwo.data.friend.local.FriendLocalDataSource
-import cord.eoeo.momentwo.data.friend.local.FriendRemoteKeyDao
+import cord.eoeo.momentwo.core.database.FriendRemoteKeyDao
 import cord.eoeo.momentwo.data.friend.remote.FriendRemoteDataSource
 import cord.eoeo.momentwo.data.friend.remote.FriendService
 import cord.eoeo.momentwo.domain.friend.FriendRepository
