+++ app/src/main/java/cord/eoeo/momentwo/domain/mapper/ProfileMapper.kt
@@ -1,7 +1,7 @@
 package cord.eoeo.momentwo.domain.mapper
 
 import cord.eoeo.momentwo.data.model.UserProfile
-import cord.eoeo.momentwo.data.profile.local.entity.ProfileEntity
+import cord.eoeo.momentwo.core.database.ProfileEntity
 import cord.eoeo.momentwo.core.model.Profile
 import cord.eoeo.momentwo.core.model.ProfileItem
 import javax.inject.Inject
