+++ core/database/src/main/java/cord/eoeo/momentwo/core/database/FriendEntity.kt
@@ -1,4 +1,4 @@
-package cord.eoeo.momentwo.data.friend.local.entity
+package cord.eoeo.momentwo.core.database
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
