+++ core/database/src/main/java/cord/eoeo/momentwo/core/database/PhotoRemoteKeyEntity.kt
@@ -1,12 +1,14 @@
-package cord.eoeo.momentwo.data.friend.local.entity
+package cord.eoeo.momentwo.core.database
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
-@Entity(tableName = "friend_key")
-data class FriendRemoteKeyEntity(
+@Entity(tableName = "photo_key")
+data class PhotoRemoteKeyEntity(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @ColumnInfo(name = "album_id") val albumId: Int,
+    @ColumnInfo(name = "sub_album_id") val subAlbumId: Int,
     @ColumnInfo(name = "last_updated") val lastUpdated: Long,
     @ColumnInfo(name = "next_cursor") val nextCursor: Int?,
 )
